refactor: pull common functionality into separate utils folder
